Refactored and renamed FileTypeValidationUtilTest.
* refactored to have "Base class" for any new FileTypeValidator
* Maven specific stuff (so far was only content of this test) moved out to MavenFileTypeValidatorTest
