+++ nexus/nexus-proxy/src/test/java/org/sonatype/nexus/proxy/maven/MavenFileTypeValidatorTest.java
@@ -16,39 +16,17 @@
  * Sonatype, Inc. Apache Maven is a trademark of the Apache Foundation. M2Eclipse is a trademark of the Eclipse Foundation.
  * All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.proxy.repository.validator;
+package org.sonatype.nexus.proxy.maven;
 
-import java.io.File;
-
-import org.codehaus.plexus.util.xml.Xpp3Dom;
-import org.junit.Assert;
 import org.junit.Test;
-import org.sonatype.nexus.configuration.model.CLocalStorage;
-import org.sonatype.nexus.configuration.model.CRepository;
-import org.sonatype.nexus.configuration.model.DefaultCRepository;
-import org.sonatype.nexus.proxy.AbstractNexusTestCase;
-import org.sonatype.nexus.proxy.ResourceStoreRequest;
-import org.sonatype.nexus.proxy.item.DefaultStorageFileItem;
-import org.sonatype.nexus.proxy.maven.MavenFileTypeValidator;
-import org.sonatype.nexus.proxy.maven.RepositoryPolicy;
-import org.sonatype.nexus.proxy.maven.maven2.M2RepositoryConfiguration;
-import org.sonatype.nexus.proxy.repository.LocalStatus;
-import org.sonatype.nexus.proxy.repository.Repository;
-import org.sonatype.nexus.proxy.storage.local.fs.FileContentLocator;
+import org.sonatype.nexus.proxy.repository.validator.AbstractFileTypeValidationUtilTest;
 
 /**
- * This test actually tests MavenFileTypeValidator. Also, it uses non-LAX XML validation! This test would pass in LAX
- * XML validation too, except one assertion, see testPom() for why's.
+ * Tests for MavenFileTypeValidator specific file types.
  */
-public class FileTypeValidationUtilTest
-    extends AbstractNexusTestCase
+public class MavenFileTypeValidatorTest
+    extends AbstractFileTypeValidationUtilTest
 {
-    private FileTypeValidatorHub getValidationUtil()
-        throws Exception
-    {
-        return lookup( FileTypeValidatorHub.class );
-    }
-
     @Test
     public void testJar()
         throws Exception
@@ -67,16 +45,20 @@ public class FileTypeValidationUtilTest
         doTest( "something/else/myapp.pom", "no-doctype-pom.xml", true );
         doTest( "something/else/myapp.pom", "simple.xml", false );
         doTest( "something/else/myapp.pom", "pom.xml", true );
+    }
+
+    @Test
+    public void testXml()
+        throws Exception
+    {
         doTest( "something/else/myapp.xml", "pom.xml", true );
+        doTest( "something/else/myapp.xml", "pom.xml", true, true );
         doTest( "something/else/myapp.xml", "simple.xml", true );
-        // Note: this last line makes this test fail for LAX XML validation
-        // In LAX mode, we do not expect from XML to have preamble but do expect it to be "plain text"
-        // Obviously, the HTML used for testing passes here.
-        // The point is, that the path "myapp.xml" is not a POM path despite this test
-        // is named as such, hence, this is actually XML validation test.
-        // And since we agreed to lax XML validation (many XMLs lacks the preamble)
-        // we cannot decide this below, unless LAX XML validation is OFF
-        doTest( "something/else/myapp.xml", "error.html", false );
+        doTest( "something/else/myapp.xml", "simple.xml", true, true );
+        // will be INVALID with XML Lax validation OFF
+        doTest( "something/else/myapp.xml", "error.html", false, false );
+        // will be VALID with XML Lax validation ON
+        doTest( "something/else/myapp.xml", "error.html", true, true );
     }
 
     @Test
@@ -140,49 +122,5 @@ public class FileTypeValidationUtilTest
         doTest( "something/else/file.jar.md5", "test.sha1", false );
         doTest( "something/else/file.jar.md5", "test.md5", true );
     }
-
-    // ==
-
-    protected Repository getDummyRepository()
-        throws Exception
-    {
-        Repository repository = lookup( Repository.class, "maven2" );
-
-        CRepository cRepo = new DefaultCRepository();
-        cRepo.setId( "test-repo" );
-        cRepo.setLocalStatus( LocalStatus.IN_SERVICE.toString() );
-        cRepo.setNotFoundCacheTTL( 1 );
-        cRepo.setLocalStorage( new CLocalStorage() );
-        cRepo.getLocalStorage().setProvider( "file" );
-        cRepo.setProviderRole( Repository.class.getName() );
-        cRepo.setProviderHint( "maven2" );
-
-        Xpp3Dom ex = new Xpp3Dom( "externalConfiguration" );
-        cRepo.setExternalConfiguration( ex );
-        M2RepositoryConfiguration extConf = new M2RepositoryConfiguration( ex );
-        extConf.setRepositoryPolicy( RepositoryPolicy.RELEASE );
-
-        repository.configure( cRepo );
-
-        return repository;
-    }
-
-    private void doTest( String expectedFileName, String testFileName, boolean expectedResult )
-        throws Exception
-    {
-        File testFile = new File( "target/test-classes/FileTypeValidationUtilTest", testFileName );
-
-        DefaultStorageFileItem file =
-            new DefaultStorageFileItem( getDummyRepository(), new ResourceStoreRequest( expectedFileName ), true, true,
-                new FileContentLocator( testFile, "this-is-neglected-in-this-test" ) );
-        file.getItemContext().put( MavenFileTypeValidator.XML_DETECTION_LAX_KEY, "false" );
-
-        boolean result = getValidationUtil().isExpectedFileType( file );
-
-        Assert.assertEquals( "File name: " + expectedFileName + " and file: " + testFileName + " match result: "
-            + result + " expected: " + expectedResult, expectedResult, result );
-    }
-
-    //
 
 }
